requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
transformers
python-multipart
pandas
